+++ webmacs/commands/global.py
@@ -17,7 +17,7 @@ import itertools
 import os
 from PyQt5.QtCore import QStringListModel, QModelIndex
 
-from . import define_command, COMMANDS
+from . import define_command, COMMANDS, register_prompt_opener_commands
 from ..minibuffer import Prompt, KEYMAP
 from ..minibuffer.prompt import PromptTableModel, PromptHistory
 from ..application import app
@@ -55,13 +55,15 @@ def quit(ctx):
     app().quit()
 
 
-@define_command("M-x", prompt=CommandsListPrompt, visible=False)
+@define_command("M-x", visible=False)
 def commands(ctx):
     """
     Prompt for a command name to execute.
     """
+    prompt = CommandsListPrompt(ctx)
+    value = ctx.minibuffer.do_prompt(prompt)
     try:
-        COMMANDS[ctx.prompt.value()](ctx)
+        COMMANDS[value](ctx)
     except KeyError:
         pass
 
@@ -258,24 +260,10 @@ class VisitedLinksPrompt(Prompt):
         "complete-empty": True,
     }
     keymap = VISITEDLINKS_KEYMAP
-
-    def enable(self, minibuffer):
-        Prompt.enable(self, minibuffer)
-        self.new_buffer = self.ctx.current_prefix_arg == (4,)
-        if self.new_buffer:
-            minibuffer.label.setText(minibuffer.label.text() + " (new buffer)")
+    value_return_index_data = True
 
     def completer_model(self):
         return VisitedLinksModel(self)
-
-    def get_buffer(self):
-        if self.new_buffer:
-            buf = create_buffer()
-            view = self.ctx.window.current_webview()
-            view.setBuffer(buf)
-        else:
-            buf = self.ctx.buffer
-        return buf
 
 
 @VISITEDLINKS_KEYMAP.define_key("C-k")
@@ -290,16 +278,11 @@ def visited_links_remove_entry(ctx):
     pinput.completer_model().remove_history_entry(selection)
 
 
-@define_command("visited-links-history", prompt=VisitedLinksPrompt)
-def visited_links_history(ctx):
-    """
-    Prompt to open a link previously visited.
-    """
-    prompt = ctx.prompt
-    index = prompt.index()
-    if index.isValid():
-        url = index.internalPointer()
-        prompt.get_buffer().load(url)
+register_prompt_opener_commands(
+    "visited-links-history",
+    VisitedLinksPrompt,
+    "Prompt to open a link previously visited",
+)
 
 
 class BookmarksModel(VisitedLinksModel):
@@ -325,12 +308,11 @@ class BookmarksPrompt(VisitedLinksPrompt):
         return BookmarksModel(self)
 
 
-@define_command("bookmark-open", prompt=BookmarksPrompt)
-def open_bookmark(ctx):
-    """
-    Prompt to open a bookmark.
-    """
-    visited_links_history(ctx)
+register_prompt_opener_commands(
+    "bookmark-open",
+    BookmarksPrompt,
+    "Prompt to open a bookmark",
+)
 
 
 class BookmarkAddPrompt(Prompt):
@@ -344,22 +326,23 @@ class BookmarkAddPrompt(Prompt):
         input.setSelection(0, len(url))
 
 
-@define_command("bookmark-add", prompt=BookmarkAddPrompt)
+@define_command("bookmark-add")
 def bookmark_add(ctx):
     """
     Create or rename a bookmark for the current url.
     """
-    minibuff = ctx.minibuffer
-    url = ctx.prompt.value()
+    prompt = BookmarkAddPrompt(ctx)
+    url = ctx.minibuffer.do_prompt(prompt)
+    if not url:
+        return
 
     otherprompt = Prompt(ctx)
     otherprompt.label = "bookmark's name: "
-    name = minibuff.do_prompt(otherprompt)
+    name = ctx.minibuffer.do_prompt(otherprompt)
 
     if name:
         app().bookmarks().set(url, name)
-        minibuff.show_info("Bookmark {} created."
-                           .format(name))
+        ctx.minibuffer.show_info("Bookmark {} created.".format(name))
 
 
 class ModesPrompt(Prompt):
@@ -368,6 +351,7 @@ class ModesPrompt(Prompt):
         "match": Prompt.FuzzyMatch,
         "complete-empty": True,
     }
+    value_return_index_data = True
 
     def completer_model(self):
         return PromptTableModel([
@@ -375,14 +359,15 @@ class ModesPrompt(Prompt):
         ])
 
 
-@define_command("buffer-set-mode", prompt=ModesPrompt)
+@define_command("buffer-set-mode")
 def buffer_set_mode(ctx):
     """
     Change the mode of the current buffer.
     """
-    index = ctx.prompt.index()
-    if index.isValid():
-        ctx.buffer.set_mode(index.internalPointer())
+    prompt = ModesPrompt(ctx)
+    mode = ctx.minibuffer.do_prompt(prompt)
+    if mode:
+        ctx.buffer.set_mode(mode)
 
 
 @define_command("send-key-down")
@@ -473,12 +458,12 @@ class VariableListPrompt(Prompt):
         return model
 
 
-@define_command("describe-variable", prompt=VariableListPrompt)
+@define_command("describe-variable")
 def describe_variable(ctx):
     """
     Prompt for a variable name to describe.
     """
-    variable = ctx.prompt.value()
+    variable = ctx.minibuffer.do_prompt(VariableListPrompt(ctx))
     if variable in variables.VARIABLES:
         buffer = create_buffer("webmacs://variable/%s" % variable)
         ctx.view.setBuffer(buffer)
@@ -489,12 +474,12 @@ class DescribeCommandsListPrompt(CommandsListPrompt):
     history = PromptHistory()
 
 
-@define_command("describe-command", prompt=DescribeCommandsListPrompt)
+@define_command("describe-command")
 def describe_command(ctx):
     """
     Prompt for a command name to describe.
     """
-    command = ctx.prompt.value()
+    command = ctx.minibuffer.do_prompt(DescribeCommandsListPrompt(ctx))
     if command in COMMANDS:
         buffer = create_buffer("webmacs://command/%s" % command)
         ctx.view.setBuffer(buffer)
@@ -521,7 +506,7 @@ class ReportCallHandler(CallHandler):
         self.prompt.minibuffer.input().setText("%s -"
                                                % self.keys_as_text())
 
-    def call(self, sender, keymap, keypress, command):
+    def call(self, ctx, keymap, keypress, command):
         self.key_presses.append(keypress)
         if not isinstance(command, str):
             command = "{}:{}".format(command.__module__,
@@ -537,6 +522,7 @@ class ReportCallHandler(CallHandler):
 
 class BindingPrompt(Prompt):
     label = "describe key: "
+    called_with = None
 
     def enable(self, minibuffer):
         self.keymap = local_keymap()
@@ -548,16 +534,21 @@ class BindingPrompt(Prompt):
         KEY_EATER.set_call_handler(self.orig_handler)
         Prompt.close(self)
 
+    def value(self):
+        return self.called_with
 
-@define_command("describe-key", prompt=BindingPrompt)
+
+@define_command("describe-key")
 def describe_binding(ctx):
     """
     Retrieve the command called by the given binding.
     """
-    url = "webmacs://key/{key}?command={command}&keymap={keymap}".format(
-        **ctx.prompt.called_with
-    )
-    ctx.view.setBuffer(create_buffer(url))
+    called_with = ctx.minibuffer.do_prompt(BindingPrompt(ctx))
+    if called_with:
+        url = "webmacs://key/{key}?command={command}&keymap={keymap}".format(
+            **called_with
+        )
+        ctx.view.setBuffer(create_buffer(url))
 
 
 @define_command("restore-session")